completely refactor payload-handler
