--- src/payload-handler/lib/create.js
@@ -1,11 +1,0 @@
-export default function (formatter, config, data) {
-  return {
-    code: '201',
-    data: formatter(data, {
-      singleResult: true
-    }),
-    headers: {
-      location: '/' + config.typeName + '/' + data.id
-    }
-  };
-}
