--- src/payload-handler/lib/destroy.js
@@ -1,6 +1,0 @@
-export default function (formatter, config, data) {
-  return {
-    code: '204',
-    data: null
-  };
-}
