--- src/payload-handler/lib/update.js
@@ -1,12 +1,0 @@
-export default function (formatter, config, data) {
-  if (data && !config.relationOnly) {
-    return {
-      code: '200',
-      data: formatter(data, config)
-    };
-  }
-  return {
-    code: '204',
-    data: null
-  };
-}
